alg dev

+++ T-Solver/src/pl/edu/agh/iet/tsp/Algorithm/Dance.java
@@ -6,9 +6,9 @@ import java.util.ArrayList;
 public class Dance {
 	
 	public ArrayList<Integer> path;
-	float duration;
+	double duration;
 	
-	public Dance(ArrayList<Integer> path, float dance_duration) {
+	public Dance(ArrayList<Integer> path, double dance_duration) {
 		this.path = path;
 		this.duration = dance_duration;
 	}
